+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_webview.dart
@@ -5,17 +5,18 @@ import 'dart:ui';
 
 import '../find_interaction/platform_find_interaction_controller.dart';
 import '../pull_to_refresh/platform_pull_to_refresh_controller.dart';
-
 import '../context_menu/context_menu.dart';
 import '../types/main.dart';
-
 import '../web_uri.dart';
 import 'in_app_webview_settings.dart';
 import 'platform_inappwebview_controller.dart';
 import '../print_job/main.dart';
+import 'platform_inappwebview_widget.dart';
+import 'platform_headless_in_app_webview.dart';
 
 ///{@template flutter_inappwebview_platform_interface.PlatformWebViewCreationParams}
-///Class that represents a WebView. Used by [InAppWebView], [HeadlessInAppWebView] and the WebView of [PlatformInAppBrowser].
+///Class that represents a WebView. Used by [PlatformInAppWebViewWidget],
+///[PlatformHeadlessInAppWebView] and the WebView of [PlatformInAppBrowser].
 ///{@endtemplate}
 class PlatformWebViewCreationParams<T> {
   final T Function(PlatformInAppWebViewController controller)?