+++ flutter_inappwebview_windows/windows/webview_environment/webview_environment_channel_delegate.h
@@ -1,8 +1,10 @@
 #ifndef FLUTTER_INAPPWEBVIEW_PLUGIN_WEBVIEW_ENVIRONMENT_CHANNEL_DELEGATE_H_
 #define FLUTTER_INAPPWEBVIEW_PLUGIN_WEBVIEW_ENVIRONMENT_CHANNEL_DELEGATE_H_
 
-#include "../types/channel_delegate.h"
 #include <flutter/method_channel.h>
+
+#include "../types/browser_process_exited_detail.h"
+#include "../types/channel_delegate.h"
 
 namespace flutter_inappwebview_plugin
 {
@@ -19,6 +21,9 @@ namespace flutter_inappwebview_plugin
     void HandleMethodCall(
       const flutter::MethodCall<flutter::EncodableValue>& method_call,
       std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result);
+
+    void onNewBrowserVersionAvailable() const;
+    void onBrowserProcessExited(std::shared_ptr<BrowserProcessExitedDetail> detail) const;
   };
 }